+++ pkgs/artisanal_widgets/test/app/artisanal_app_test.dart
@@ -1,8 +1,8 @@
 import 'package:artisanal/style.dart' show Layout;
+import 'package:artisanal/widgets.dart' as w;
 import 'package:artisanal/tui.dart'
     show BackgroundColorMsg, KeyMsg, View, WindowSizeMsg;
 import 'package:artisanal/terminal.dart' show Key, KeyType;
-import 'package:artisanal_widgets/artisanal_widgets.dart' as w;
 import 'package:test/test.dart';
 
 void main() {